Rebuild of neighbor list, topology, and parameters are now correct.
I changed a bunch of stuff to make sure everything is right. Before, we would run out of memory due to a small bug. However, the parameters and topology could get out of sync with the neighbor list. I switched to an n-squared neighbor list and rebuild on every step to make sure everything is right. The groundwork is now laid to refactor everything into a proper system and force field object now.