Began implementing CMM in new abstraction
Finished SPCfw in new abstraction

Began changing metadata stored on each Term so that the parameterizer can compute all of the needed params at the beginning of a frame and store the parameters in the storage container. This is important both because it provides caching for terms that use the same parameters and because it will eventually allow us allocation-free execution of most the computational graph.