Can accumulate gradients across multiple evaluations
Need to come up with a better way of resetting the force field than rebuilding it entirely. Like have the force field specify how to be rebuilt. Also, in the long term, there should be a way to evaluate an entire dataset in parallel by specifying "segments" in the topology which result in disjoint sets of pairs that will be still evaluated simultaneously. Will have to change kernels to return multiple energy values, rather than one but that should be fine.